add logining

+++ src/Xyzies.SSO.Identity/Xyzies.SSO.Identity.UserMigration/Services/Migrations/MigrationService.cs
@@ -296,7 +296,7 @@ namespace Xyzies.SSO.Identity.UserMigration.Services.Migrations
                         return;
                     }
                     PrepeareUserProperties(entity, roles, statuses, companyBranches);
-                    await _azureClient.PostUser(entity.Adapt<AzureUser>());
+                    var insertedUser = await _azureClient.PostUser(entity.Adapt<AzureUser>());
 
                     if (!string.IsNullOrWhiteSpace(entity.State))
                     {
@@ -308,7 +308,7 @@ namespace Xyzies.SSO.Identity.UserMigration.Services.Migrations
                         await _locationService.SetState(entity.City);
                     }
 
-                    _userService.UpdateUserInCache(entity.Adapt<AzureUser>());
+                    _userService.UpdateUserInCache(insertedUser);
                     _logger.LogInformation($"New user, {entity.Name} {entity.LastName} {entity.Role ?? "NULL ROLE!!!"}");
                 }
 
@@ -345,7 +345,13 @@ namespace Xyzies.SSO.Identity.UserMigration.Services.Migrations
                 {
                     entity.Email = string.IsNullOrEmpty(_migrationPostfix) ? entity.Email : entity.Email + _migrationPostfix;
                     var existUser = await _userService.GetUserBy(u => u.SignInNames.FirstOrDefault(name => name.Type == "emailAddress")?.Value.ToLower() == entity.Email.ToLower());
+                    if(existUser == null)
+                    {
+                        _logger.LogInformation($"User {entity.Email} was deleted from Cable Portal, but was not found in Azure");
+                        return;
+                    }
                     await _userService.DeleteUserByIdAsync(existUser.ObjectId);
+                   _logger.LogInformation($"User {entity.Email} was deleted");
                 }
             }
             catch (Exception ex)
